Linux: Handle protected_regular in inplace writes
Linux kernel 4.19 and systemd 241 introduced a new filesystem
protection feature called "protected_regular":
https://www.kernel.org/doc/Documentation/sysctl/fs.txt

From the documentation:

> When set to "1" don't allow O_CREAT open on regular files that we
> don't own in world writable sticky directories, unless they are
> owned by the owner of the directory.

> When set to "2" it also applies to group writable sticky directories."

This may end up in "rsync --inplace" failing to write to some files, even
if run as root. A simple "cp -a" succeeds, though, as rsync uses the
"O_CREAT" option, while cp does not. This change works around the issue
by detecting the failures and trying again without "O_CREAT". The idea
is that these types of failures are rare, so we don't want to change the
default logic.

Added a simple test case. As we need chown to trigger the failure, this
test is skipped if not running as root.