Add option --log-after to log after moving file into place
This mode is useful when a process is monitoring the log for
post-processing of transferred files.

With --log-after in local mode, both sender and receiver log to
the same log file, so it require --log-file with absolute path.

We add %o to the default log format, so it will be easy to tell
the logs of the sender from the logs of the receiver:

2023/02/14 14:40:25 [559755] building file list
2023/02/14 14:40:25 [559755] send >f+++++++++ foo
2023/02/14 14:40:25 [559757] recv >f+++++++++ foo
2023/02/14 14:40:25 [559755] sent 111 bytes  received 35 bytes  292.00 bytes/sec
2023/02/14 14:40:25 [559755] total size is 4  speedup is 0.03