Fix runtime AVX2 detection for checksum
The way GCC implements function-multiversioning is to generate a
dispatcher function, thus it could not be used when a function is not
built by GCC (e.g. a direct .S file, in the situation of AVX2 checksum).

Use function-multiversioning to declare a wrapper that will calls the
appropriate checksum generatator function when AVX2 is present and do
nothing when not present.

Signed-off-by: Icenowy Zheng <uwu@icenowy.me>