Add --chmod-dest option for handling --fake-super permissions
When --chmod is used with --fake-super, the permissions specified
by chmod are written both to the destination file, and the fake
super attributes, which can in some cases make it difficult to
provide access to a fake super backup.

The --chmod-dest option when used with --fake-super allows fake
super to back up the source permissions, while applying different
destination permissions to the backup up file.