harden systemd service with safe defaults
This commit adds hardening to the systemd service with a variety of
knobs designed to restrict damage that rsync can do as root, while
making sure that no file transfers are interrupted. For example,
ProtectHome= is still off, and CapabilityBoundingSet is still unlimited
to allow for chown/chmod to work unimpeded.

The rsync@.service has some additional options to restrict all network
access and only use the socket on stdin. An equivalent change was done
in rsync.service to only allow TCP, but full network isolation is not
possible for that unit.