Add lsattr/chattr support to --xattrs
This is like `bsdtar --fflags`.

Repurposing `sys_llistxattr` allows cross-platform interoperability,
such as for a transfer from Linux to Mac then from Mac to Linux.
It's also simpler by not affecting the rest of the codebase.

There is no new cmdline option because #165 wants this to work with just `-avX`.

Closes: #165
Closes: #508